requirements.txt
Django>=2.2.9,<2.3.0
django-guardian>=2.1.0,<2.2.0
django-jet>=1.0.8,<1.1.0
djangorestframework>=3.11.0,<3.12.0
djangorestframework-simplejwt>=4.4.0,<4.5.0
django-cors-headers>=3.2.1,<3.3.0
flake8>=3.7.9,<3.8.0
Pillow>=6.2.1,<6.3.1
protobuf>=3.11.2,<3.12.0
psycopg2>=2.8.3,<2.9.0
PyJWT>=1.7.1,<1.8.0
pytz==2019.3
